part_2_ai_basics/chapter_07_intro_ai_llm/requirements.txt
# requirements.txt
openai==2.9.0
python-dotenv==1.2.1
anthropic==0.75.0
requests==2.32.5
google-generativeai==0.8.5
pandas==2.3.3